Cleaned up some of the internal policy bits code. Working on restructuring some of the detaching, especially w.r.t. policy contexts (as encoded in one of the policy properties). Still dealing with a major IRET issue in the kernel.

+++ deps/dr/x86/instr.h
@@ -1764,7 +1764,7 @@ struct _instr_t {
     app_pc  translation;
 
     IF_GRANARY_ELSE(uint16_t, uint)    opcode;
-    IF_GRANARY(uint8_t granary_policy; )
+    IF_GRANARY(uint16_t granary_policy; )
     IF_GRANARY(uint8_t granary_flags; )
 
 #ifdef X64
