+++ granary/dynamorio.h
@@ -653,7 +653,7 @@ struct _instr_t {
     uint length;
     app_pc translation;
     uint16_t opcode;
-    uint8_t granary_policy; // code cache policy for granary
+    uint16_t granary_policy; // code cache policy for granary
     uint8_t granary_flags; // flags, e.g. delay, don't mangle, etc.
     byte rip_rel_pos;
     byte num_dsts;
